% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstats.R
\name{bootstats}
\alias{bootstats}
\title{Obtain mean and median with bootstrapped confidence intervals.}
\usage{
bootstats(x, p.level = 0.95, nrep = NULL, exact.thrs = 5, smooth = TRUE,
  return_resamples = FALSE)
}
\arguments{
\item{x}{A vector of numerical values.}
}
\description{
This function returns the bootstrapped mean and median, 
and the respective confidence intervals, of the supplied numeric vector.
}
\examples{
bootstats(1:4)

bootstats(1:4, nrep=9999)

## Simple smooth function based on jitter()
bootstats(1:5, smooth=function(b=b) jitter(b))

## Alternative entropy based smooth function
x <- round(rnorm(15, 10, 2))

entropy <- function(x, base=2) {
    freqs <- table(x) / length(x)
    -sum(freqs * log(freqs, base=base))
}

H5 <- entropy(x, base=5)

bootstats(x, 
  smooth=function(b=b, x=x, H5=H5) {
      b + rnorm(length(b), 0, H5/sqrt(length(x)))
  })

bootstats(x)
}
\keyword{univar}
