% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstats.R
\name{multisubsets}
\alias{multisubsets}
\title{Multisubsets of cardinality k}
\usage{
multisubsets(v, k = length(v))
}
\arguments{
\item{v}{A vector to sample from}

\item{k}{The size of the subsamples}
}
\description{
All possible ways to sample k elements from a given set with replacement.
}
\examples{

multisubsets(1:3, 4)

multisubsets(1:4, 3)

multisubsets(c("red", "blue", "green"))
}
