% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstats.R
\name{print.bootlist}
\alias{print.bootlist}
\title{Print method for class bootlist.}
\usage{
\method{print}{bootlist}(x, ...)
}
\description{
Print method for class bootlist.
}
\examples{
x <- c(1, 3, 4, 2, 8)
bst <- bootstats(x, nrep=5, return_resamples=TRUE)
bst
str(bst)
bst$resamples
}
